✨ Make next explicit flow control
tldr; You must call `next` in middleware unless you want to stop the
exectution.

Previously, `next` was called implicitly by middleware composition, if
it was not called from the middleware itself.

While this behavior is convinient for some cases, it takes away the
possibility to skip the subsequent middleware execution for the purpose
of caching or any other early exit. This behavior is consistent with
other frameworks, like Express or Koa.