🏷️ Fix multiple handlers w/ServiceImplementation
Adding multiple handlers worked correctly for inlined handlers and type
helper with key access, but not with typed service implementation:

```typescript
const handler = ServiceImplementation<IService>['method']
const serviceImpl = ServiceImplementation<IService>

app.use(Service, {
  method: [
    call => call, // inline def: OK
    handler, // helper type direct: OK
    serviceImpl.method, // helper type, whole service: ERR
  ]
}
```

```
Type '((call: ProtoCatCall<...>, next: NextFn) => any) | ((call: ProtoCatCall<...>, next: NextFn) => any)[]'
    is not assignable to type '(call: ProtoCatCall<...>, next: NextFn) => any
```
TS cannot recognize in this context that `serviceImpl.method` is not an
array, which would create a nested array.

And it is correct assumption ofc. This could be solved with generics,
which would make the helper terribly complex to use (would require a
lambda), or by splitting the helper into standard versions with single
handlers and extended with possibility to have an array of handlers.

The latter was implemented and tested.