✅ Update tests for middleware initial data
After update in prev commit, initial metadata set in mdw after next
function did not reach the client for sync handlers for bidi and
server-streaming calls.

The existing test is now split in two parts:
 - initial metadata from mdw
 - initial metadata from mdw after next function (skipped for the two
   call types)

To make this work, the flush would need to be delayed to the point when
all middlewares are finsihed. Doable, but skipping for now.